part 2 of the porting

+++ src/main/java/com/telepathicgrunt/repurposedstructures/misc/MobSpawnerObj.java
@@ -1,8 +1,8 @@
 package com.telepathicgrunt.repurposedstructures.misc;
 
 import com.google.gson.annotations.Expose;
-import net.minecraft.entity.EntityType;
-import net.minecraft.util.ResourceLocation;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.entity.EntityType;
 import net.minecraftforge.registries.ForgeRegistries;
 
 public class MobSpawnerObj {
@@ -10,13 +10,13 @@ public class MobSpawnerObj {
     public String name;
 
     @Expose
-    public int weight;
+    public float weight;
 
     public transient EntityType<?> entityType;
 
-    public MobSpawnerObj(String name, double weight){
+    public MobSpawnerObj(String name, float weight) {
         this.name = name;
-        this.weight = (int)weight;
+        this.weight = weight;
     }
 
     public void setEntityType() throws Exception {
